refactor(PN-11896): wip
