+++ packages/pn-personafisica-webapp/vite.config.mts
@@ -34,7 +34,7 @@ export default defineConfig(({ mode }) => {
       open: true,
       proxy: {
         '^/auth/.*': {
-          target: 'https://login.dev.notifichedigitali.it',
+          target: 'https://login.test.notifichedigitali.it',
           changeOrigin: true,
         },
       },
